Propaga el borrado de un usuario al borrado del perfil relacionado

+++ app/controllers/user.controller.js
@@ -79,26 +79,29 @@ exports.findAll = (req, res) => {
     });
 };
 
-// Find a single User with an id
-exports.findOne = (req, res) => {
-  const id = req.params.id;
+exports._findOne = async (id) => {
+  try {
+    const user = await User.findByPk(id);
 
-  User.findByPk(id)
-    .then(data => {
-      if (data) {
-        res.send(data);
-      } else {
-        res.status(404).send({
-          message: `Cannot find User with id=${id}.`
-        });
-      }
-    })
-    .catch(err => {
-      console.log(err.message);
-      res.status(500).send({
-        message: `Error retrieving User with id=${id}.`
-      });
-    });
+    return {
+      status: user ? 200 : 404,
+      payload: user ? user : { message: `Cannot find User with id=${id}.` }
+    };
+  } catch (err) {
+    console.log(err.message);
+    return {
+      status: 500,
+      payload: { message: `Error retrieving User with id=${id}.` }
+    };
+  }
+}
+
+// Find a single User with an id
+exports.findOne = async (req, res) => {
+  const id = req.params.id;
+  const response = await this._findOne(id);
+
+  res.status(response.status).send(response.payload);
 };
 
 // Update a User by the id in the request
@@ -106,7 +109,7 @@ exports.update = (req, res) => {
   const id = req.params.id;
 
   User.update(req.body, {
-    where: { id: id }
+    where: { id }
   })
     .then(num => {
       if (num == 1) {
@@ -127,28 +130,42 @@ exports.update = (req, res) => {
 };
 
 // Delete a User with the specified id in the request
-exports.delete = (req, res) => {
+exports.delete = async (req, res) => {
   const id = req.params.id;
+  const userResponse = await this._findOne(id);
+  
+  if (userResponse.status == 200) {
+    const user = userResponse.payload;
+    const profileId = user.profileId;
+    const profileResponse = await ProfileController._delete(profileId);
 
-  User.destroy({
-    where: { id: id }
-  })
-    .then(num => {
-      if (num == 1) {
-        res.send({
-          message: 'User was deleted successfully.'
+    if (profileResponse.status == 200) {
+      User.destroy({
+        where: { id }
+      })
+        .then(num => {
+          if (num == 1) {
+            res.send({
+              message: 'User was deleted successfully.'
+            });
+          } else {
+            res.send({
+              message: `Cannot delete User with id=${id}. Maybe User was not found or req.body is empty!`
+            });
+          }
+        })
+        .catch(err => {
+          res.status(500).send({
+            message: `Error deleting User with id=${id}`
+          });
         });
-      } else {
-        res.send({
-          message: `Cannot delete User with id=${id}. Maybe User was not found or req.body is empty!`
-        });
-      }
-    })
-    .catch(err => {
-      res.status(500).send({
-        message: `Error deleting User with id=${id}`
-      });
-    });
+      return;
+    }
+
+    return res.status(profileResponse.status).send(profileResponse.payload);
+  }
+
+  res.status(userResponse.status).send(userResponse.payload);
 };
 
 // // Delete all Users from the database
